# apps/web/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Medicly API",
    "version": "1.0.0",
    "description": "Medicly MVP HTTP API. Cookie session: medicly_sid."
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "medicly_sid"
      }
    },
...
  },
  "paths": {
...
    "/v1/notifications/{id}/read": {
      "post": {
        "operationId": "postNotificationRead",
        "tags": [
          "notifications"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/patients/me/favourites/{doctorId}": {
      "post": {
        "operationId": "postFavourite",
        "tags": [
          "patients"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "doctorId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok",
                    "isFavourite"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "isFavourite": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteFavourite",
        "tags": [
          "patients"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "doctorId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok",
                    "isFavourite"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "isFavourite": {
                      "type": "boolean",
                      "enum": [
                        false
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/patients/me/recently-viewed/{doctorId}": {
      "post": {
        "operationId": "postRecentlyViewed",
        "tags": [
          "patients"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "doctorId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/reviews": {
      "post": {
        "operationId": "postReview",
        "tags": [
          "patients"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "appointmentId",
                  "rating"
                ],
                "properties": {
                  "appointmentId": {
                    "type": "string"
                  },
                  "rating": {
                    "minimum": 1,
                    "maximum": 5,
                    "type": "integer"
                  },
                  "text": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "review"
                  ],
                  "properties": {
                    "review": {
                      "type": "object",
                      "required": [
                        "id",
                        "appointmentId",
                        "doctorId",
                        "rating",
                        "text",
                        "createdAt"
                      ],
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "appointmentId": {
                          "type": "string"
                        },
                        "doctorId": {
                          "type": "string"
                        },
                        "rating": {
                          "type": "number"
                        },
                        "text": {
                          "type": "string",
                          "nullable": true
                        },
                        "createdAt": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/api",
      "description": "Vite proxy / same-origin /api"
    }
  ],
...
}